// app/src/types/align.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * A qué línea se lleva cada elemento.
 */
export type Alignment = "left" | "center_x" | "right" | "top" | "middle" | "bottom";

/**
 * En qué eje se reparte el espacio.
 */
export type Spread = "horizontal" | "vertical";